cleaneup controllers
