+++ src/main/java/ua/com/owu/controller/MainController.java
@@ -9,17 +9,22 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
-import ua.com.owu.models.Account;
-import ua.com.owu.models.Admin;
-import ua.com.owu.models.Role;
+import ua.com.owu.models.*;
 import ua.com.owu.service.AccountService.AccountService;
 import ua.com.owu.service.MailService;
 
 import ua.com.owu.utils.AccountEditor;
+import ua.com.owu.utils.TokenUtils;
+import ua.com.owu.utils.UserValidator;
 
 import javax.mail.MessagingException;
+import java.io.IOException;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Controller
 public class MainController {
@@ -31,12 +36,18 @@ public class MainController {
     @Autowired
     private MailService mailService;
 
-
     @Autowired
     AccountService accountService;
 
     @Autowired
     AccountEditor accountEditor;
+
+    @Autowired
+    UserValidator userValidator;
+
+    @Autowired
+    TokenUtils tokenUtils;
+
 
     @GetMapping("/")
     public String index() {
@@ -57,7 +68,7 @@ public class MainController {
         return "index";
     }
 
-
+    //temporary methods to  create  admin
     @GetMapping("/createAdmin")
     public String createAdmin() {
         Admin admin = new Admin(Role.ROLE_ADMIN, "admin", "admin", "as@as");
@@ -68,56 +79,86 @@ public class MainController {
 
     @PostMapping("/ok")
     public String ok(Model model) {
-
-
         return "ok";
     }
 
-    @GetMapping("/books/input")
-    public String bookInput() {
-        return "";
+    //AdminController
+    @GetMapping("/admin/page")
+    String adminPage(Model model) {
+
+        List<Account> manager = accountService.findByAccountType("manager");
+        Stream<Account> stream = manager.stream();
+        List<Account> collect = stream.filter(account -> account.isAccountNonLocked() == false).collect(Collectors.toList());
+        model.addAttribute("manager", collect);
+        return "adminT";
     }
 
-
-    @GetMapping("/times")
-    public String times() {
-        return "index";
-    }
-
-    @PostMapping("/times")
-    public String timesSave(
-            @RequestParam String date, @RequestParam String start, @RequestParam String end
+    @GetMapping("/admin/active/manager/id/{id}")
+    String confirm(
+            @PathVariable int id
     ) {
-
-
-        DateTime startDay = DateTime.parse(date);
-        DateTime endDay = startDay.plusHours(24);
-        DateTime startTime = startDay.plusHours(Integer.parseInt(start));
-        DateTime endTime = startDay.plusHours(Integer.parseInt(end));
-        Interval interval = new Interval(startTime, endTime);
-        Interval interval1 = new Interval(startDay, endDay);
-
-        System.out.println(startDay);
-        System.out.println(endDay);
-        System.out.println(startTime);
-        System.out.println(endTime);
-        System.out.println(interval1.gap(interval));
-        return "redirect:/times";
+        Account managerAccount = accountService.findbyId(id);
+        managerAccount.setAccountNonLocked(true);
+        accountService.save(managerAccount);
+        return "redirect:/admin/page";
     }
 
-    @PostMapping("/sentMail")
-    public String sentMail(
-            @RequestParam String email,
-            @RequestParam String subject,
-            @RequestParam String message
+    //USERcONTROLLER
 
-    ) {
+
+    @PostMapping("/save")
+    public String save(User user,
+                       BindingResult bindingResult,
+                       Model model
+    ) throws IOException {
+        String username = user.getUsername();
+        userValidator.validate(user, bindingResult);
+        user.setToken(tokenUtils.generateToken());
+        if (bindingResult.hasErrors()) {
+            String errorMessage = "";
+            List<ObjectError> allErrors = bindingResult.getAllErrors();
+            for (ObjectError allError : allErrors) {
+                String code = allError.getCode();
+                errorMessage += " " + environment.getProperty(code);
+            }
+            model.addAttribute("error", errorMessage);
+            return "index";
+        }
         try {
-            mailService.sendSimpleMessage(email, subject, message);
+            mailService.sendConfirmMessage(user.getEmail(), user);
         } catch (MessagingException e) {
             e.printStackTrace();
         }
+        accountEditor.setValue(user);
+        accountService.save(user);
+        return "redirect:/";
+    }
 
+
+
+    @GetMapping("/confirm/{token}")
+    public String accontConfirm(@PathVariable String token) {
+        Account byToken = accountService.findByToken(token);
+        if (byToken != null) {
+            byToken.setToken(null);
+            byToken.setEnabled(true);
+            accountService.save(byToken);
+            System.out.println(byToken.getUsername() + " is approveed!");
+            return "login";
+        } else {
+            System.out.println("there is  no tokens  like  that!");
+            return "index";
+        }
+    }
+
+    //MANAGERCONTROLLER
+
+    @PostMapping("/save/manager")
+    public String manager(Manager manager) {
+        accountEditor.setValue(manager);
+        manager.setRole(Role.ROLE_MANAGER);
+        manager.setAccountNonLocked(false);
+        accountService.save(manager);
         return "redirect:/";
     }
